Add role field for assistant messages (#89)

+++ examples/sync/assistants/threads.py
@@ -2,14 +2,19 @@
 
 from __future__ import annotations
 
+import pathlib
+
 from yandex_cloud_ml_sdk import YCloudML
+
+PATH = pathlib.Path(__file__)
+NAME = f'example-{PATH.parent.name}-{PATH.name}'
 
 
 def main() -> None:
     sdk = YCloudML(folder_id='b1ghsjum2v37c2un8h64')
     sdk.setup_default_logging()
 
-    thread = sdk.threads.create(name='foo', ttl_days=5, expiration_policy="static")
+    thread = sdk.threads.create(name=NAME, ttl_days=5, expiration_policy="static")
     print(f"new {thread=}")
 
     second = sdk.threads.get(thread.id)
@@ -17,9 +22,13 @@ def main() -> None:
     second.update(ttl_days=9)
     print(f"with updated epiration config, {second=}")
 
-    message = thread.write("content")
-    message2 = second.write("content2")
-    print(f"hey, we just writed {message=} and {message2} into the thread")
+    # You could pass string
+    thread.write("content")
+    # {"text": str, "role": str} dict
+    message2 = second.write({"text": "content2", "role": "ASSISTANT"})
+    # or any object which have .text and .role attributes, like a Message object
+    # from assistants
+    second.write(message2)
 
     print("and now we could read it:")
     for message in thread:
@@ -29,8 +38,9 @@ def main() -> None:
         print(f"    {message.status.name=}\n")
 
     for thread in sdk.threads.list():
-        print(f"deleting thread {thread=}")
-        thread.delete()
+        if thread.name == NAME:
+            print(f"deleting thread {thread=}")
+            thread.delete()
 
 
 if __name__ == '__main__':
